+++ models/timeline/DataProvider.php
@@ -5,20 +5,20 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\debug\components;
+namespace yii\debug\models\timeline;
 
 use yii\data\ArrayDataProvider;
 use yii\debug\panels\TimelinePanel;
 
 /**
- * TimelineDataProvider implements a data provider based on a data array.
+ * DataProvider implements a data provider based on a data array.
  *
  * @property array $rulers This property is read-only.
  *
  * @author Dmitriy Bashkarev <dmitriy@bashkarev.com>
- * @since 2.0.7
+ * @since 2.0.8
  */
-class TimelineDataProvider extends ArrayDataProvider
+class DataProvider extends ArrayDataProvider
 {
     /**
      * @var TimelinePanel
@@ -27,7 +27,7 @@ class TimelineDataProvider extends ArrayDataProvider
 
 
     /**
-     * TimelineDataProvider constructor.
+     * DataProvider constructor.
      * @param TimelinePanel $panel
      * @param array $config
      */
@@ -142,6 +142,26 @@ class TimelineDataProvider extends ArrayDataProvider
             $data[$ms] = $ms / $percent;
         }
         return $data;
+    }
+
+    /**
+     * ```php
+     * [
+     *   0 => string, memory usage (MB)
+     *   1 => float, Y position (percent)
+     * ]
+     * @param array $model
+     * @return array|null
+     */
+    public function getMemory($model)
+    {
+        if (empty($model['memory'])) {
+            return null;
+        }
+        return [
+            sprintf('%.2f MB', $model['memory'] / 1048576),
+            $model['memory'] / ($this->panel->memory / 100)
+        ];
     }
 
 }