+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 debug extension Change Log
 2.0.8 under development
 -----------------------
 
+- Enh #34: Added memory graph to timeline panel (bashkarev)
 - Enh #179: Increased request time logging accuracy and precision (samdark)
 - Enh #174: Added routing panel (bashkarev, samdark)
 - Enh #185: Added meta tag to prevent indexing of debug by search engines in case it's exposed (aminkt, samdark)